# src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: TaskMaster API
  version: 1.0.0
  description: A RESTful API for managing tasks in a task management application.
servers:
  - url: http://localhost:3000
paths:
  /tasks:
    get:
      summary: Retrieve all tasks
      operationId: getAllTasks
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /tasks/{id}:
    get:
      summary: Retrieve a task by ID
      operationId: getTaskById
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the task to retrieve
          schema:
            type: string
      responses:
        '200':
          description: A task object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
    put:
      summary: Update a task by ID
      operationId: updateTask
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the task to update
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
    delete:
      summary: Delete a task by ID
      operationId: deleteTask
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the task to delete
          schema:
            type: string
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean
    TaskInput:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean
          default: false